+++ CountlyContentBuilderInternal.h
@@ -16,6 +16,10 @@ NS_ASSUME_NONNULL_BEGIN
 @property (nonatomic, assign) NSTimeInterval zoneTimerInterval;
 @property (nonatomic) ContentCallback contentCallback;
 @property (nonatomic, assign) WebViewDisplayOption webViewDisplayOption;
+@property (nonatomic, assign) BOOL enableContentReloadOnStall;
+@property (nonatomic, assign) NSTimeInterval contentReloadOnStallTimeout; // seconds
+@property (nonatomic, assign) BOOL disableZoom;
+@property (nonatomic, copy, nullable) ContentURLHandler contentURLHandler;
 @property (nonatomic, assign) int contentInitialDelay;
 
 + (instancetype)sharedInstance;