+++ CountlyContentConfig.m
@@ -11,6 +11,10 @@
 @property (nonatomic) ContentCallback contentCallback;
 @property (nonatomic) NSUInteger zoneTimerInterval;
 @property (nonatomic) WebViewDisplayOption webViewDisplayOption;
+@property (nonatomic) BOOL contentReloadOnStallEnabled;
+@property (nonatomic) NSUInteger contentReloadOnStallTimeoutMs;
+@property (nonatomic) BOOL zoomDisabled;
+@property (nonatomic, copy) ContentURLHandler contentURLHandler;
 #endif
 @end
 
@@ -20,8 +24,11 @@
 {
     if (self = [super init])
     {
+#if (TARGET_OS_IOS)
+        _contentReloadOnStallTimeoutMs = 1000; // default 1 second
+#endif
     }
-    
+
     return self;
 }
 
@@ -57,6 +64,46 @@
 - (WebViewDisplayOption)getWebViewDisplayOption;
 {
     return _webViewDisplayOption;
+}
+
+- (void)enableContentReloadOnStall
+{
+    _contentReloadOnStallEnabled = YES;
+}
+
+- (BOOL)getEnableContentReloadOnStall
+{
+    return _contentReloadOnStallEnabled;
+}
+
+- (void)setContentReloadOnStallTimeout:(NSUInteger)milliseconds
+{
+    _contentReloadOnStallTimeoutMs = milliseconds;
+}
+
+- (NSUInteger)getContentReloadOnStallTimeout
+{
+    return _contentReloadOnStallTimeoutMs;
+}
+
+- (void)disableZoom
+{
+    _zoomDisabled = YES;
+}
+
+- (BOOL)getDisableZoom
+{
+    return _zoomDisabled;
+}
+
+- (void)setContentURLHandler:(ContentURLHandler)handler
+{
+    _contentURLHandler = handler;
+}
+
+- (ContentURLHandler)getContentURLHandler
+{
+    return _contentURLHandler;
 }
 #endif